@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <PublishWithAspNetCoreTargetManifest>false</PublishWithAspNetCoreTargetManifest>
   </PropertyGroup>
 
@@ -19,7 +19,6 @@
 
   <ItemGroup>
     <PackageReference Include="BitMiracle.LibTiff.NET" Version="2.4.649" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.5" />
   </ItemGroup>
 
   <ItemGroup>
@@ -36,10 +35,10 @@
     <PackageReference Include="Syncfusion.Compression.Net.Core" Version="*" />
     <PackageReference Include="Syncfusion.OfficeChart.Net.Core" Version="*" />
     <PackageReference Include="Syncfusion.EJ2.WordEditor.AspNet.Core" Version="*" />
-	<PackageReference Include="Syncfusion.EJ2.SpellChecker.AspNet.Core" Version="*" />
-	<PackageReference Include="System.Drawing.Common" Version="4.*" />	
-	<PackageReference Include="System.Linq.Dynamic.Core" Version="1.0.9.2" />
-  <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.*" />
+    <PackageReference Include="Syncfusion.EJ2.SpellChecker.AspNet.Core" Version="*" />
+    <PackageReference Include="System.Drawing.Common" Version="4.*" />
+	<PackageReference Include="Syncfusion.Licensing" Version="*" />
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.*" />
   </ItemGroup>
 
 </Project>
