@@ -1,35 +1,24 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.IO;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.ResponseCompression;
+using Newtonsoft.Json.Serialization;
+using Syncfusion.EJ2.SpellChecker;
+using Newtonsoft.Json;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Data.Edm;
-using Microsoft.AspNet.OData.Builder;
-using Microsoft.AspNet.OData.Extensions;
-using Microsoft.AspNetCore.Routing;
-using Microsoft.OData.UriParser;
-using Microsoft.AspNetCore.Cors.Infrastructure;
-using EJ2APIServices.Models;
-using EJ2APIServices.Controllers;
-using Syncfusion.EJ2.SpellChecker;
-using System.IO;
-using Newtonsoft.Json;
-
+using System.Collections.Generic;
+using Microsoft.Extensions.Hosting;
 
 namespace EJ2APIServices
 {
     public class Startup
     {
-        private string _contentRootPath = "";
         internal static string path;
 
-        public Startup(IConfiguration configuration, IHostingEnvironment env)
+        public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
             var builder = new ConfigurationBuilder()
                 .SetBasePath(env.ContentRootPath)
@@ -38,82 +27,83 @@ namespace EJ2APIServices
                 .AddEnvironmentVariables();
 
             Configuration = builder.Build();
-            _contentRootPath = env.ContentRootPath;
 
             path = Configuration["SPELLCHECK_DICTIONARY_PATH"];
             string jsonFileName = Configuration["SPELLCHECK_JSON_FILENAME"];
             //check the spell check dictionary path environment variable value and assign default data folder
             //if it is null.
-            path = string.IsNullOrEmpty(path) ? Path.Combine(env.ContentRootPath, "App_Data") : Path.Combine(env.ContentRootPath, path);
+            path = string.IsNullOrEmpty(path) ? Path.Combine(env.ContentRootPath, "Data") : Path.Combine(env.ContentRootPath, path);
             //Set the default spellcheck.json file if the json filename is empty.
             jsonFileName = string.IsNullOrEmpty(jsonFileName) ? Path.Combine(path, "spellcheck.json") : Path.Combine(path, jsonFileName);
-            if (System.IO.File.Exists(jsonFileName))
+            if (File.Exists(jsonFileName))
             {
-                string jsonImport = System.IO.File.ReadAllText(jsonFileName);
+                string jsonImport = File.ReadAllText(jsonFileName);
                 List<DictionaryData> spellChecks = JsonConvert.DeserializeObject<List<DictionaryData>>(jsonImport);
                 List<DictionaryData> spellDictCollection = new List<DictionaryData>();
                 string personalDictPath = null;
                 //construct the dictionary file path using customer provided path and dictionary name
-                foreach (var spellCheck in spellChecks)
+                if (spellChecks != null)
                 {
-                    spellDictCollection.Add(new DictionaryData(spellCheck.LanguadeID, Path.Combine(path, spellCheck.DictionaryPath), Path.Combine(path, spellCheck.AffixPath)));
-                    personalDictPath = Path.Combine(path, spellCheck.PersonalDictPath);
+                    foreach (var spellCheck in spellChecks)
+                    {
+                        spellDictCollection.Add(new DictionaryData(spellCheck.LanguadeID, Path.Combine(path, spellCheck.DictionaryPath), Path.Combine(path, spellCheck.AffixPath)));
+                        personalDictPath = Path.Combine(path, spellCheck.PersonalDictPath);
+                    }
                 }
                 SpellChecker.InitializeDictionaries(spellDictCollection, personalDictPath, 3);
             }
         }
 
         public IConfiguration Configuration { get; }
-
+        readonly string MyAllowSpecificOrigins = "MyPolicy";
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddOData();
-            services.AddMvc().AddJsonOptions(x => {
-                x.SerializerSettings.ContractResolver = null;
+            services.AddControllers();
+            services.AddMemoryCache();
+            services.AddControllers().AddJsonOptions(options =>
+            {
+                options.JsonSerializerOptions.PropertyNamingPolicy = null;
             });
 
             services.AddCors(options =>
             {
-                options.AddPolicy("AllowAllOrigins", builder =>
+                options.AddPolicy(MyAllowSpecificOrigins,
+                builder =>
                 {
                     builder.AllowAnyOrigin()
-                    .AllowAnyMethod()
-                    .AllowAnyHeader();
+                      .AllowAnyMethod()
+                      .AllowAnyHeader();
                 });
             });
-            string connection = Configuration.GetConnectionString("Test");
-            if (connection.Contains("%CONTENTROOTPATH%"))
-            {
-                connection = connection.Replace("%CONTENTROOTPATH%", _contentRootPath);
-
-            }
-            services.AddEntityFrameworkNpgsql().AddDbContext<EEJ2SERVICEEJ2WEBSERVICESSRCAPP_DATADIAGRAMMDFContext>(options => options.UseNpgsql(connection));
-            // Add framework services.
-            services.AddMvc();
+            services.Configure<GzipCompressionProviderOptions>(options => options.Level = System.IO.Compression.CompressionLevel.Optimal);
+            services.AddResponseCompression();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            app.UseDeveloperExceptionPage();
-            app.UseCors("AllowAllOrigins");
+            //Register Syncfusion license
+            string licenseKey = string.Empty;
+            Syncfusion.Licensing.SyncfusionLicenseProvider.RegisterLicense(licenseKey);
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-
-            app.UseMvc(b =>
+            else
             {
-                b.Count().Filter().OrderBy().Expand().Select().MaxTop(null);
+                app.UseHsts();
+            }
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseAuthorization();
+            app.UseCors(MyAllowSpecificOrigins);
+            app.UseResponseCompression();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers().RequireCors("MyPolicy");
             });
-
         }
-
-    }
-
-    public class ServerPath
-    {
-        public static string MapPath = "";
     }
 }
